Update README.md
Used Python to build a model to predict the number of confirmed COVID-19 cases. Predicted the infection rate, and the number of deaths in the country of Zambia based on collected data from the early stages of the pandemic and created graphs and models to show findings based on calculations.